# implement faster floating-point `isless`