@ViralBShah 2021-01-05
@stev47 Can you provide the before/after benchmark times? Just for curious onlookers who may not check out the branch and run the script themselves.