@stev47 2021-01-05
> @stev47 Can you provide the before/after benchmark times? Just for curious onlookers who may not check out the branch and run the script themselves.

I've updated the benchmarks. The times got stripped out by accident, since `git commit` removes lines starting with `#` in the commit message.