@musm 2021-01-06
Is the only reason we are *not* using `<`, is to satisfy the ordering  `totalOrder` ?   Simply piggy backing off of `<`, but accounting for the ordering, results in close performance to this PR (near identical for arrays without NaNs in most cases). The disadvantage is it's not as fast in cases with mixed NaN's, but the advantage is that it's simpler to implement and requires fewer LoC.

```julia
@inline function isless(a::T, b::T) where T<:Union{Float16,Float32,Float64}
    (isnan(a) || isnan(b)) && return !isnan(a)
    return a < b
end
```


On my computer: Timings of this change in the only case where the timings are significantly different (i.e. arrays containing mixed NaN's):
```julia
a = [rand((rand(), NaN)) for _ in 1:1000000];
@btime sort($a, lt=(a,b)->isless(a,b));

# 42% faster  (this PR)
# 31% faster  (using isless with < directly)
```
Otherwise the changes within 0-3%.

I'm not necesarily advocating to change the PR to this, but just food for thought.