@stev47 2021-01-06
> Is the only reason we are _not_ using `<`, is to satisfy the ordering `totalOrder` ? Simply piggy backing off of `<`, but accounting for the ordering, results in close performance to this PR (near identical for arrays without NaNs in most cases).

NaNs are not the only difference: e.g. `isless(-0., 0.)` is true, while `-0. < 0.` is not. Non-canonical encodings are also treated differently.