@vtjnash 2021-01-10
Any particular knowledge of why LLVM doesn't already emit this for AMD64? I have a mild concern that this could be more expensive for other platforms, since moving values from fp to int registers could be slow (while this benchmark simply avoids ever having them in fp registers), and that this requires doing 64-bit operations without the benefit of the 64-bit double hardware.