@stev47 2021-01-10
> Any particular knowledge of why LLVM doesn't already emit this for AMD64?

There shouldn't be anything special about amd64 here, llvm just doesn't seem to generate efficient code for the current suboptimal C implementation of  `fpislt` in `src/runtime_instrinsics.c:838`:  The equivalence to the signed-integer interpretation is not obvious, also llvm didn't manage to simplify that NaN-check.

> since moving values from fp to int registers could be slow

The current C implementation already does that using union access (technically undefined behaviour in C).

> while this benchmark simply avoids ever having them in fp registers

I'm not entirely convinced this is true since the NaN check always seems to be done in fp registers.
But even if you force `@noinline` on the comparator (and thereby force passing float registers to the function) there is a benefit:
```julia
a = rand(1000000);
@btime sort($a, lt=@noinline (a,b)->isless(a,b));
# before: 188.884 ms (2 allocations: 7.63 MiB)
# after: 162.796 ms (2 allocations: 7.63 MiB)
```
Alternatively check
```julia
isless($(Ref(0.))[], $(Ref(0.))[])
# before: 3.399 ns (0 allocations: 0 bytes)
# after: 2.671 ns (0 allocations: 0 bytes)
```

Generated code by `@code_native isless(0., 0.)`, before:
```asm
	.text
	vmovq	%xmm0, %rax
	vmovq	%xmm1, %rcx
	testq	%rax, %rax
	sets	%dl
	setns	%sil
	cmpq	%rcx, %rax
	seta	%al
	setl	%cl
	andb	%dl, %al
	andb	%sil, %cl
	orb	%al, %cl
	vucomisd	%xmm1, %xmm0
	setnp	%dl
	andb	%cl, %dl
	vucomisd	%xmm1, %xmm1
	setp	%cl
	vucomisd	%xmm0, %xmm0
	setnp	%al
	andb	%cl, %al
	orb	%dl, %al
	retq
```
after:
```asm
	.text
	vucomisd	%xmm1, %xmm0
	jp	L56
	vmovq	%xmm0, %rax
	movabsq	$9223372036854775807, %rcx # imm = 0x7FFFFFFFFFFFFFFF
	movq	%rax, %rdx
	xorq	%rcx, %rdx
	testq	%rax, %rax
	cmovnsq	%rax, %rdx
	vmovq	%xmm1, %rax
	xorq	%rax, %rcx
	testq	%rax, %rax
	cmovnsq	%rax, %rcx
	cmpq	%rcx, %rdx
	setl	%al
	retq
L56:
	vucomisd	%xmm0, %xmm0
	setnp	%al
	retq
```